refractor: add subfolders to screens
