+++ lib/screens/scaffolds/cafe_menu/cafe_menu_scaffold.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:staugustinechsnewapp/screens/main/cafe_menu_screen.dart';
+import 'package:staugustinechsnewapp/screens/main/cafe_menu/cafe_menu_screen.dart';
 import 'package:staugustinechsnewapp/utilities/cafe_menu/cafe_menu_bloc.dart';
 import 'package:staugustinechsnewapp/widgets/reusable/custom_snackbar.dart';
