+++ lib/screens/screen_controller.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:staugustinechsnewapp/screens/scaffolds/cafe_menu_scaffold.dart';
-import 'package:staugustinechsnewapp/screens/scaffolds/home_scaffold.dart';
-import 'package:staugustinechsnewapp/screens/scaffolds/join_clubs_scaffold.dart';
+import 'package:staugustinechsnewapp/screens/scaffolds/cafe_menu/cafe_menu_scaffold.dart';
+import 'package:staugustinechsnewapp/screens/scaffolds/home/home_scaffold.dart';
 import 'package:staugustinechsnewapp/screens/scaffolds/login_scaffold.dart';
-import 'package:staugustinechsnewapp/screens/scaffolds/profile_scaffold.dart';
-import 'package:staugustinechsnewapp/screens/scaffolds/settings_scaffold.dart';
-import 'package:staugustinechsnewapp/screens/scaffolds/socials_scaffold.dart';
-import 'package:staugustinechsnewapp/screens/scaffolds/song_request_scaffold.dart';
+import 'package:staugustinechsnewapp/screens/scaffolds/profile/profile_scaffold.dart';
+import 'package:staugustinechsnewapp/screens/scaffolds/profile/settings_scaffold.dart';
+import 'package:staugustinechsnewapp/screens/scaffolds/socials/join_clubs_scaffold.dart';
+import 'package:staugustinechsnewapp/screens/scaffolds/socials/socials_scaffold.dart';
+import 'package:staugustinechsnewapp/screens/scaffolds/song_requests/song_request_scaffold.dart';
 import 'package:staugustinechsnewapp/styles.dart';
 import 'package:staugustinechsnewapp/utilities/auth/auth_bloc.dart';
 import 'package:staugustinechsnewapp/utilities/navigation/enav_index_conversion.dart';
